+++ src/com/amd/docker/dockerNodes.groovy
@@ -1,7 +1,7 @@
 /* ************************************************************************
  * Copyright 2018-2019 Advanced Micro Devices, Inc.
  * ************************************************************************ */
-package com.amd.docker.dockerNodes
+package com.amd.docker
 
 import com.amd.project.*
 import com.amd.docker.rocDocker
