+++ src/com/amd/docker/rocDocker.groovy
@@ -1,7 +1,8 @@
 /* ************************************************************************
  * Copyright 2018-2019 Advanced Micro Devices, Inc.
  * ************************************************************************ */
-package com.amd.docker.rocDocker
+
+package com.amd.docker
 import java.nio.file.Path;
 
 // Docker related variables gathered together to reduce parameter bloat on function calls
